Refactor WaitForImageTest and update WaitForTextTests
